FIX: CPP03. a copy constructor in child class calls the copy constructor in parent. the same for assignation operator in child class

+++ CPP_MODULE_03/ex02/ClapTrap.cpp
@@ -104,7 +104,7 @@ void	ClapTrap::beRepaired(unsigned int amount)
 	std::cout << " is repaired with Energy_points ";
 	std::cout << amount << std::endl;
 
-	if (this->Energy_points + amount > this->Max_energy_points)
+	if (this->Energy_points + amount > (unsigned int)this->Max_energy_points)
 		this->Energy_points = this->Max_energy_points;
 	else
 		this->Energy_points += amount;
